ALS-6330: (WIP) Integration test refactor, variant explorer fixes.
