+++ data/src/main/java/edu/harvard/hms/dbmi/avillach/hpds/data/genotype/VariantMask.java
@@ -33,4 +33,6 @@ public interface VariantMask {
     }
 
     Set<Integer> patientMaskToPatientIdSet(List<String> patientIds);
+
+    boolean isEmpty();
 }
