--- etl/src/test/resources/bucketIndexBySampleTest_vcfIndex.tsv
@@ -1,2 +1,0 @@
-filename	chromosome	annotated	gzip	sample_ids	patient_ids	sample_relationship	related_sample_ids
-src/test/resources/BucketIndexTestInput.vcf	ALL	1	0	HG00096,HG00097,HG00099,HG00100,HG00101,HG00102,HG00103,HG00105,HG00106,HG00107	1,2,3,4,5,6,7,8,9,10
