+++ integration-test/src/main/java/edu/harvard/hms/dbmi/avillach/hpds/test/Main.java
@@ -1,0 +1,7 @@
+package edu.harvard.hms.dbmi.avillach.hpds.test;
+
+public class Main {
+    public static void main(String[] args) {
+        System.out.println("Hello world!");
+    }
+}
