+++ processing/src/main/java/edu/harvard/hms/dbmi/avillach/hpds/processing/GenomicProcessorNodeImpl.java
@@ -98,6 +98,7 @@ public class GenomicProcessorNodeImpl implements GenomicProcessor {
                     intersectionOfInfoFilters = VariantIndex.empty();
                 }
             }
+
             // todo: handle empty getVariantInfoFilters()
 
             // add filteredIdSet for patients who have matching variants, heterozygous or homozygous for now.
@@ -112,17 +113,7 @@ public class GenomicProcessorNodeImpl implements GenomicProcessor {
             VariantBucketHolder<VariableVariantMasks> variantMasksVariantBucketHolder = new VariantBucketHolder<>();
             if (!distributableQuery.getRequiredFields().isEmpty() ) {
                 for (String variantSpec : distributableQuery.getRequiredFields()) {
-                    VariantMask patientsForVariantSpec = getIdSetForVariantSpecCategoryFilter(new String[]{"0/1", "1/1"}, variantSpec, variantMasksVariantBucketHolder);
-                    if (patientMask == null) {
-                        patientMask = patientsForVariantSpec;
-                    } else {
-                        patientMask = patientMask.intersection(patientsForVariantSpec);
-                    }
-                }
-            }
-            if (!distributableQuery.getCategoryFilters().isEmpty()) {
-                for (Map.Entry<String, String[]> categoryFilterEntry : distributableQuery.getCategoryFilters().entrySet()) {
-                    VariantMask patientsForVariantSpec = getIdSetForVariantSpecCategoryFilter(categoryFilterEntry.getValue(), categoryFilterEntry.getKey(), null);
+                    VariantMask patientsForVariantSpec = getVariantIndexesForSpec(new String[]{"0/1", "1/1"}, variantSpec);
                     if (patientMask == null) {
                         patientMask = patientsForVariantSpec;
                     } else {
@@ -135,6 +126,36 @@ public class GenomicProcessorNodeImpl implements GenomicProcessor {
         }
         return createMaskForPatientSet(distributableQuery.getPatientIds());
         /* END OF VARIANT INFO FILTER HANDLING */
+    }
+
+    private VariantMask getVariantIndexesForSpec(String[] zygosities, String variantSpec) {
+        List<VariableVariantMasks> masksForDbSnpSpec = variantService.getMasksForDbSnpSpec(variantSpec);
+        VariantMask variantMask = VariantMask.emptyInstance();
+
+        for (VariableVariantMasks masks : masksForDbSnpSpec) {
+            for (String zygosity : zygosities) {
+                if(zygosity.equals(HOMOZYGOUS_REFERENCE)) {
+                    /*// todo: implement this -- difficult with sparse variants of unknown length
+                    VariantMask homozygousReferenceBitmask = calculateIndiscriminateBitmask(masks);
+                    for(int x = 2;x<homozygousReferenceBitmask.bitLength()-2;x++) {
+                        homozygousReferenceBitmask = homozygousReferenceBitmask.flipBit(x);
+                    }
+                    variantBitmasks.add(homozygousReferenceBitmask);*/
+                } else if(masks.heterozygousMask != null && zygosity.equals(HETEROZYGOUS_VARIANT)) {
+                    variantMask = variantMask.union(masks.heterozygousMask);
+                } else if(masks.homozygousMask != null && zygosity.equals(HOMOZYGOUS_VARIANT)) {
+                    variantMask = variantMask.union(masks.homozygousMask);
+                } else if(zygosity.equals("")) {
+                    if (masks.homozygousMask != null) {
+                        variantMask = variantMask.union(masks.homozygousMask);
+                    }
+                    if (masks.heterozygousMask != null) {
+                        variantMask = variantMask.union(masks.heterozygousMask);
+                    }
+                }
+            }
+        }
+        return variantMask;
     }
 
     @Override
@@ -241,39 +262,30 @@ public class GenomicProcessorNodeImpl implements GenomicProcessor {
                 unionOfInfoFilters = addVariantsForInfoFilter(unionOfInfoFilters, filter);
             }
 
-            HashSet<Integer> allPatients = new HashSet<>(
-                    Arrays.stream(variantService.getPatientIds())
-                            .map((id) -> {
-                                return Integer.parseInt(id.trim());
-                            })
-                            .collect(Collectors.toList()));
-            Set<Integer> patientSubset = Sets.intersection(query.getPatientIds(), allPatients);
+            VariantMask patientMask = runGetPatientMask(query);
 
             // If we have all patients then no variants would be filtered, so no need to do further processing
-            if(patientSubset.size()==variantService.getPatientIds().length) {
+            /*if(patientSubset.size()==variantService.getPatientIds().length) {
                 log.info("query selects all patient IDs, returning....");
                 return unionOfInfoFilters.mapToVariantSpec(variantService.getVariantIndex());
-            }
-
-            VariantMask patientMasks = createMaskForPatientSet(patientSubset);
+            }*/
 
             Set<String> unionOfInfoFiltersVariantSpecs = unionOfInfoFilters.mapToVariantSpec(variantService.getVariantIndex());
-            Collection<String> variantsInScope = variantService.filterVariantSetForPatientSet(unionOfInfoFiltersVariantSpecs, new ArrayList<>(patientSubset));
+            Collection<String> variantsInScope = variantService.filterVariantSetForPatientSet(unionOfInfoFiltersVariantSpecs, patientMaskToPatientIdSet(patientMask));
 
             //NC - this is the original variant filtering, which checks the patient mask from each variant against the patient mask from the query
             if(variantsInScope.size()<100000) {
                 ConcurrentSkipListSet<String> variantsWithPatients = new ConcurrentSkipListSet<>();
                 variantsInScope.parallelStream().forEach(variantKey -> {
                     variantService.getMasks(variantKey, new VariantBucketHolder<>()).ifPresent(masks -> {
-                        // todo: implement for variant explorer
-                        /*if ( masks.heterozygousMask != null && masks.heterozygousMask.intersection(patientMasks).bitCount()>4) {
+                        if ( masks.heterozygousMask != null && !masks.heterozygousMask.intersection(patientMask).isEmpty()) {
                             variantsWithPatients.add(variantKey);
-                        } else if ( masks.homozygousMask != null && masks.homozygousMask.intersection(patientMasks).bitCount()>4) {
+                        } else if ( masks.homozygousMask != null && !masks.homozygousMask.intersection(patientMask).isEmpty()) {
                             variantsWithPatients.add(variantKey);
-                        } else if ( masks.heterozygousNoCallMask != null && masks.heterozygousNoCallMask.intersection(patientMasks).bitCount()>4) {
+                        } else if ( masks.heterozygousNoCallMask != null && !masks.heterozygousNoCallMask.intersection(patientMask).isEmpty()) {
                             //so heterozygous no calls we want, homozygous no calls we don't
                             variantsWithPatients.add(variantKey);
-                        }*/
+                        }
                     });
                 });
                 return variantsWithPatients;
