+++ processing/src/main/java/edu/harvard/hms/dbmi/avillach/hpds/processing/VariantIndexCache.java
@@ -80,7 +80,7 @@ public class VariantIndexCache {
                 int x = 0;
                 for(Integer variantIndex : variantIndexIntArray) {
                     // todo: shouldn't this be greater than or equal to 0? 0 is a valid index
-                    if (variantIndex > 0) {
+                    if (variantIndex >= 0) {
                         variantIndexArray[variantIndex] = true;
                     }
                 }
