+++ processing/src/test/java/edu/harvard/hms/dbmi/avillach/hpds/processing/AbstractProcessorTest.java
@@ -43,7 +43,8 @@ public class AbstractProcessorTest {
                 null,
                 infoStores,
                 null,
-                genomicProcessor
+                genomicProcessor,
+                ""
         );
     }
