+++ service/src/main/java/edu/harvard/hms/dbmi/avillach/hpds/service/HpdsApplication.java
@@ -2,7 +2,9 @@ package edu.harvard.hms.dbmi.avillach.hpds.service;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.support.PropertySourcesPlaceholderConfigurer;
 
 @SpringBootApplication
 @ComponentScan("edu.harvard.hms.dbmi.avillach.hpds")
@@ -10,6 +12,11 @@ public class HpdsApplication {
 
     public static void main(String[] args) {
         SpringApplication.run(HpdsApplication.class, args);
+    }
+
+    @Bean
+    public static PropertySourcesPlaceholderConfigurer propertySourcesPlaceholderConfigurer() {
+        return new PropertySourcesPlaceholderConfigurer();
     }
 
 }
