+++ service/src/main/resources/application-integration-test.properties
@@ -2,5 +2,6 @@ SMALL_JOB_LIMIT = 100
 SMALL_TASK_THREADS = 1
 LARGE_TASK_THREADS = 1
 
-hpds.genomicProcessor.impl=localDistributed
-HPDS_GENOMIC_DATA_DIRECTORY=/opt/local/hpds/phs0000946/
+hpds.genomicProcessor.impl=local
+HPDS_GENOMIC_DATA_DIRECTORY=target/
+HPDS_DATA_DIRECTORY=target/test-classes/phenotypic/
